+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [3.3.3] - 2021-02-04
 ### Added
 - Can now define a property path for the region on `PhoneNumber` constraint
 - New option to specify default region for serialization